Fixes #3389: Move r27561 to v1.7 branch (retain full FQDN hostnames)
---svn-pre-commit-ignore-below---

r27561 [[BR]]
If the nodename is an IP address, we need to retain the full name (even if keep_fqdn is false) so that the ssh tree spawn can proceed.

cmr:v1.7


r27583 [[BR]]
Add a function to test if a string is in the form of an IP address - doesn't test for validity of the address

r27586 [[BR]]
Strip the domain info from the hostname if that option is specified, protecting IP address-based names

r27618 [[BR]]
Add a function to return the aliases (based on IP addrs) for the current node

r27619 [[BR]]
Add an MCA param to retain all aliases based on IP addrs for node names so that procs can look them up by interface, if desired. If the param is set, pass aliases around to all daemons and procs for local use

r27620 [[BR]]
Per Jeff's suggestion, set the default answer when asking for IP aliases in case we don't find any