Fixes #3448: Move r27740 to v1.7 branch
---svn-pre-commit-ignore-below---

r27740 [[BR]]
Fix a bug in the uGNI btl that could cause certain descriptor callbacks to be called twice.

There was a race condition in the eager get protocol where the RDMA complete message could be received before the local completion of the SMSG message that started the eager get protocol.

cmr:v1.7