Fixes #3232: 2D string array handling broken in Fortran MPI_COMM_SPAWN_MULTIPLE
---svn-pre-commit-ignore-below---

r28164 [[BR]]
Ugh.  Work around an issue with memory hooks and the change from one big
library to multiple libraries that are implicitly sucked into the executable
as a dependency of libmpi.  The initialize hook isn't visible to libc on some
linux distributions when it's in libopal and libopal isn't explicity linked
into the executable.  The fix is to have a duplicate initialize hook in
libmpi as well as libopal.  *sigh*.