Fixes #3537: Upgrade hwloc from 1.5.1 --> 1.5.2
---svn-pre-commit-ignore-below---

r28178 [[BR]]
Update hwloc from 1.5.1 to 1.5.2.  Re-enable hwloc PCI support by
default, since it will now use libpciaccess (if available).

r28212 [[BR]]
Re-apply r28040 from Eugene: a post-hwloc release fix for Solaris
binding.  This fix was included in the upstream 1.6 series, but not
the upstream 1.5 series, and was therefore missed when we brought
1.5.2 to OMPI.

r28215 [[BR]]
Include <hwloc.h> so that the symbols in this file are subject to the
<hwloc/rename.h> renaming.

r28218 [[BR]]
Update the post-hwloc-1.5.2 patch list.

r28224 [[BR]]
Remove a duplicate variable declaration.