Fixes #3359: Move OPAL hotel class to v1.7 series (rollup)
---svn-pre-commit-ignore-below---

r27067 [[BR]]
Add new "opal_hotel" class.  Abstractly speaking, this class does the
following:

 * Provides a fixed number of resource slots (i.e., "hotel rooms").
 * Allows one thing to occupy a resource slot at a time (i.e., each
   hotel room can have an occupant check in to that room).
 * Resource slots can be vacated at any time (i.e., occupants can
   voluntarily check out of their hotel room).
 * Resource slots can be occupied for a specific maximum amount of
   time.  If that time expires, the occupant is forcibly evicted and
   the upper layer is notified via (libevent) callback (i.e., the maid
   will kick an occupant of out of their room when their reservation
   is over).

This class can be to be used for things like retransmission schemes
for unreliable transports.  For example, a message sent on an
unreliable transport can be checked in to a hotel room.  If an ACK for
that message is received, the message can be checked out.  But if the
ACK is never received, the message will eventually be evicted from its
room and the upper layer will be notified that the message failed to
check out in time (i.e., that an ACK for that message was not received
in time).

Code using this class is currently being developed off-trunk, but will
be coming to SVN soon.

r27068 [[BR]]
Gah -- meant to make these changes before committing to SVN.  :-\

Hide some struct declarations in the .c file to emphasize that they
are not part of the public opal_hotel interface.

r27075 [[BR]]
Gah!!  We inlined some of the functionality, so we need these structs
to be defined.  Put comments in there indicating that they're private
and should not be used by public consumers.

r27462 [[BR]]
Fix backwards asserts in the OPAL hotel code.

r27555 [[BR]]
There's no point in having a separate opal_hotel_finalize() function
-- just move that functionality into the hotel destructor.