Fix for CSCug82600: on the receive side, ensure that a frag is in a hotel room before we attempt to check it out.
Specifically: there's a case where an ACK arrives for a frag that is
not in a hotel room (so don't check it out!).

When a sent frag is evicted from a hotel room (i.e., the frag was
sent, but we did not receive an ACK back within the specified
timeout), if there are no send WQEs available to re-send it
immediately, we queue up that frag to be re-sent.  But if the ACK for
that frag arrives while it is still queued up to be re-sent, that frag
is not in the hotel room that the endpoint send window says it is.

The moral of the story: when a good ACK arrives, only check the
corresponding frag out of its hotel if the frag's state_flags have
FRAG_IN_HOTEL set.