Fix for CSCug82600:
When resending a message, we reserve several WQEs for the PML.  If we
don't do this, we may completely exhaust SQ WQEs with resends, and
then a btl.send downcall will return OMPI_ERR_OUT_OF_RESOURCE to the
PML.  The PML will queue up the frag, but since the SQ was full of
resends, none of them will trigger an upcall to the PML when they
complete -- so the PML will never be notified that it should try to
send its previously-deferred message again.  Deadlock!

So we reserve a few WQEs here to ensure that that will never happen.

That being said, a *better* solution will be to move all fragmenting
down here to the BTL, but that's a larger overhaul and will come in a
future version.