Fix for CSCuh04264
When a usnic BTL module is created, set its device_context->async_fd
in a libevent.  If anything is ever available to read off that fd,
call the PML error callback routine (which currently aborts the job,
but do intelligent things -- i.e., abort the job -- if the PML error
callback ever actually returns).

Note that I list all the async events that usnic_verbs.ko currently
can send, to include one that is not yet in RHEL 6
/usr/include/infiniband/verbs.h (and added a configure.m4 check for
that one).  Also had to add the infrastructure to get the PML error
callback function pointer.